cursor/requirements.txt
playwright
beautifulsoup4
pydantic
python-dotenv 